Added file size method to NSString category Rmeoved bandaid patch on rooted paths
git-svn-id: http://opensource.utr-software.com/source/connection/trunk@479 30aebe3c-6c0d-0410-bd16-8e6c7f3d8b55

+++ CKTransferRecord.m
@@ -347,7 +347,7 @@ NSString *CKTransferRecordProgressChangedNotification = @"CKTransferRecordProgre
 {
 	myTransferred += length;
 	NSTimeInterval now = [NSDate timeIntervalSinceReferenceDate];
-	[self setSpeed:length / (now - myLastTransferTime)];
+	[self setSpeed:((double)length) / (now - myLastTransferTime)];
 	[self willChangeValueForKey:@"speed"];
 	myLastTransferTime = now;
 	[self didChangeValueForKey:@"speed"];
